Refactor SSL checks
